Move tests/beacon/state_machines/conftest to upper dir
